US-3487 [IMP] SFTP auto import dev NOT TESTED

+++ bin/addons/msf_tools/automated_import_job.py
@@ -48,13 +48,13 @@ def all_files_under(path, startswith=False):
         break # don't parse children
     return res
 
-def get_oldest_filename(job, ftp_connec=None):
+def get_oldest_filename(job, ftp_connec=None, sftp=None):
     '''
     Get the oldest file in local or on FTP server
     '''
     if not job.import_id.ftp_source_ok:
         return min(all_files_under(job.import_id.src_path, job.import_id.function_id.startswith), key=os.path.getmtime)
-    else:
+    elif job.import_id.ftp_protocol == 'ftp':
         files = []
         ftp_connec.dir(job.import_id.src_path, files.append)
         file_names = []
@@ -70,9 +70,18 @@ def get_oldest_filename(job, ftp_connec=None):
             dt = time.strptime(dt, '%Y%m%d%H%M%S') # '20180228170748'
             res.append((dt, file))
         return min(res, key=lambda x:x[1])[1] if res else False
+    elif job.import_id.ftp_protocol == 'sftp':
+        latest = 0
+        latestfile = False
+        with sftp.cd(job.import_id.src_path):
+            for fileattr in sftp.listdir_attr():
+                if fileattr.st_mtime > latest:
+                    latest = fileattr.st_mtime
+                    latestfile = fileattr.filename
+        return os.path.join(job.import_id.src_path, latestfile) if latestfile else False
 
 
-def get_file_content(file, from_ftp=False, ftp_connec=None):
+def get_file_content(file, from_ftp=False, ftp_connec=None, sftp=None):
     '''
     get file content from local of FTP
     If ftp_connec is given then we try to retrieve line from FTP server
@@ -81,13 +90,19 @@ def get_file_content(file, from_ftp=False, ftp_connec=None):
         ch.write('%s\n' % line)
     if not from_ftp:
         return open(file).read()
-    else:
+    elif ftp_connec:
         ch = StringIO()
         ftp_connec.retrlines('RETR %s' % file, add_line)
         return ch.getvalue()
+    elif sftp:
+        tmp_file_path = os.path.join(tempfile.gettempdir(), os.path.basename(file))
+        sftp.get(file, tmp_file_path)
+        with open(tmp_file_path, 'r') as fich:
+            return fich.read()
+    return False
 
 
-def move_to_process_path(import_brw, ftp_connec, file, success):
+def move_to_process_path(import_brw, ftp_connec, sftp, file, success):
     """
     Move the file `file` from `src_path` to `dest_path`
     :return: return True
@@ -97,19 +112,20 @@ def move_to_process_path(import_brw, ftp_connec, file, success):
 
     dest_on_ftp = import_brw.ftp_dest_ok if success else import_brw.ftp_dest_fail_ok
 
-    if import_brw.ftp_source_ok and dest_on_ftp:
+    ############################################## FTP #########################################################
+    if import_brw.ftp_source_ok and dest_on_ftp and import_brw.ftp_protocol == 'ftp':
         # from FTP to FTP
         rep = ftp_connec.rename(srcname, destname)
         if not rep.startswith('2'):
             raise osv.except_osv(_('Error'), ('Unable to move file to destination location on FTP server'))
-    elif not import_brw.ftp_source_ok and dest_on_ftp:
+    elif not import_brw.ftp_source_ok and dest_on_ftp and import_brw.ftp_protocol == 'ftp':
         # from local to FTP
         rep = ftp_connec.storbinary('STOR %s' % destname, open(srcname, 'rb'))
         if not rep.startswith('2'):
             raise osv.except_osv(_('Error'), ('Unable to move local file to destination location on FTP server'))
         else:
             os.remove(srcname)
-    elif import_brw.ftp_source_ok and not dest_on_ftp:
+    elif import_brw.ftp_source_ok and not dest_on_ftp and import_brw.ftp_protocol == 'ftp':
         # from FTP to local
         rep = ''
         with open(destname, 'wb') as f:
@@ -122,6 +138,28 @@ def move_to_process_path(import_brw, ftp_connec, file, success):
             rep2 = ftp_connec.delete(srcname)
             if not rep2.startswith('2'):
                 raise osv.except_osv(_('Error'), ('Unable to remove remote file on FTP server'))
+    ############################################## SFTP #########################################################
+    elif import_brw.ftp_source_ok and dest_on_ftp and import_brw.ftp_protocol == 'sftp':
+        # from FTP to FTP
+        try:
+            sftp.rename(srcname, destname)
+        except:
+            raise osv.except_osv(_('Error'), ('Unable to move file to destination location %s on SFTP server') % destname)
+    elif not import_brw.ftp_source_ok and dest_on_ftp and import_brw.ftp_protocol == 'sftp':
+        # from local to FTP
+        try:
+            sftp.put(srcname, destname)
+            os.remove(srcname)
+        except:
+            raise osv.except_osv(_('Error'), ('Unable to move local file to destination location on SFTP server'))
+    elif import_brw.ftp_source_ok and not dest_on_ftp and import_brw.ftp_protocol == 'sftp':
+        # from FTP to local
+        try:
+            sftp.get(srcname, destname, preserve_mtime=True)
+            sftp.delete(srcname)
+        except:
+            raise osv.except_osv(_('Error'), ('Unable to move remote file to local destination location on SFTP server'))
+    ############################################### LOCAL ####################################################
     else:
         # from local to local
         shutil.move(srcname, destname)
@@ -252,9 +290,12 @@ class automated_import_job(osv.osv):
             filename = False
 
             ftp_connec = None
-            if job.import_id.ftp_ok:
-                context.update({'no_raise_if_ok': True})
+            sftp = None
+            context.update({'no_raise_if_ok': True})
+            if job.import_id.ftp_ok and job.import_id.ftp_protocol == 'ftp':
                 ftp_connec = self.pool.get('automated.import').ftp_test_connection(cr, uid, job.import_id.id, context=context)
+            elif job.import_id.ftp_ok and job.import_id.ftp_protocol == 'sftp':
+                sftp = self.pool.get('automated.import').sftp_test_connection(cr, uid, job.import_id.id, context=context)
 
             try:
                 for path in [('src_path', 'r', 'ftp_source_ok'), ('dest_path', 'w', 'ftp_dest_ok'), ('dest_path_failure', 'w', 'ftp_dest_fail_ok'), ('report_path', 'w', 'ftp_report_ok')]:
@@ -268,12 +309,12 @@ class automated_import_job(osv.osv):
 
             if not job.file_to_import:
                 try:
-                    oldest_file = get_oldest_filename(job, ftp_connec)
+                    oldest_file = get_oldest_filename(job, ftp_connec, sftp)
                     if not oldest_file:
                         raise ValueError()
                     filename = os.path.split(oldest_file)[1]
-                    md5 = hashlib.md5(get_file_content(oldest_file, job.import_id.ftp_source_ok, ftp_connec)).hexdigest()
-                    data64 = base64.encodestring(get_file_content(oldest_file, job.import_id.ftp_source_ok, ftp_connec))
+                    md5 = hashlib.md5(get_file_content(oldest_file, job.import_id.ftp_source_ok, ftp_connec, sftp)).hexdigest()
+                    data64 = base64.encodestring(get_file_content(oldest_file, job.import_id.ftp_source_ok, ftp_connec, sftp))
                 except ValueError:
                     no_file = True
 
@@ -282,7 +323,7 @@ class automated_import_job(osv.osv):
                         error = _('No file to import in %s !') % job.import_id.src_path
                     elif md5 and self.search_exist(cr, uid, [('import_id', '=', job.import_id.id), ('file_sum', '=', md5)], context=context):
                         error = _('A file with same checksum has been already imported !')
-                        move_to_process_path(job.import_id, ftp_connec, filename, success=False)
+                        move_to_process_path(job.import_id, ftp_connec, sftp, filename, success=False)
                         self.infolog(cr, uid, _('%s :: Import file (%s) moved to destination path') % (job.import_id.name, filename))
 
                 if error:
@@ -329,22 +370,27 @@ class automated_import_job(osv.osv):
             error_message = []
             state = 'done'
             try:
-                if job.import_id.ftp_source_ok:
+                if job.import_id.ftp_source_ok and job.import_id.ftp_protocol == 'ftp':
                     prefix = '%s_' % filename.split('.')[0]
                     suffix = '.xls' if self.pool.get('stock.picking').get_import_filetype(cr, uid, filename) == 'excel' else '.xml' 
                     temp_file = tempfile.NamedTemporaryFile(delete=False, prefix=prefix, suffix=suffix)
                     ftp_connec.retrbinary('RETR %s' % oldest_file, temp_file.write)
                     temp_file.close()
                     oldest_file = temp_file.name
+                elif job.import_id.ftp_source_ok and job.import_id.ftp_protocol == 'sftp':
+                    tmp_dest_path = os.path.join(tempfile.gettempdir(), filename)
+                    sftp.get(oldest_file, tmp_dest_path)
+                    oldest_file = tmp_dest_path
+
                 processed, rejected, headers = getattr(
                     self.pool.get(job.import_id.function_id.model_id.model),
                     job.import_id.function_id.method_to_call
                 )(cr, uid, oldest_file)
                 if processed:
-                    nb_processed = self.generate_file_report(cr, uid, job, processed, headers, ftp_connec=ftp_connec)
+                    nb_processed = self.generate_file_report(cr, uid, job, processed, headers, ftp_connec=ftp_connec, sftp=sftp)
 
                 if rejected:
-                    nb_rejected = self.generate_file_report(cr, uid, job, rejected, headers, rejected=True, ftp_connec=ftp_connec)
+                    nb_rejected = self.generate_file_report(cr, uid, job, rejected, headers, rejected=True, ftp_connec=ftp_connec, sftp=sftp)
                     state = 'error'
                     for resjected_line in rejected:
                         line_message = _('Line %s: ' % resjected_line[0])
@@ -364,7 +410,7 @@ class automated_import_job(osv.osv):
                     'state': state,
                 }, context=context)
                 is_success = True if not rejected else False
-                move_to_process_path(job.import_id, ftp_connec, filename, success=is_success)
+                move_to_process_path(job.import_id, ftp_connec, sftp, filename, success=is_success)
                 self.infolog(cr, uid, _('%s :: Import file (%s) moved to destination path') % (job.import_id.name, filename))
             except Exception as e:
                 cr.rollback()
@@ -380,7 +426,7 @@ class automated_import_job(osv.osv):
                     'file_to_import': data64,
                     'state': 'error',
                 }, context=context)
-                move_to_process_path(job.import_id, ftp_connec, filename, success=False)
+                move_to_process_path(job.import_id, ftp_connec, sftp, filename, success=False)
                 self.infolog(cr, uid, _('%s :: Import file (%s) moved to destination path') % (job.import_id.name, filename))
 
         return {
@@ -394,7 +440,7 @@ class automated_import_job(osv.osv):
         }
 
 
-    def generate_file_report(self, cr, uid, job_brw, data_lines, headers, rejected=False, ftp_connec=None):
+    def generate_file_report(self, cr, uid, job_brw, data_lines, headers, rejected=False, ftp_connec=None, sftp=None):
         """
         Create a csv file that contains the processed lines and put this csv file
         on the report_path directory and attach it to the automated.import.job.
@@ -417,7 +463,7 @@ class automated_import_job(osv.osv):
         delimiter = ','
         quotechar = '"'
         on_ftp = job_brw.import_id.ftp_report_ok
-        assert not on_ftp or (on_ftp and ftp_connec), _('FTP connection issue')
+        assert not on_ftp or (on_ftp and ftp_connec) or (on_ftp and sftp), _('FTP connection issue')
 
         csvfile = tempfile.NamedTemporaryFile(mode='wb', delete=False) if on_ftp else open(pth_filename, 'wb')
         if on_ftp:
@@ -434,11 +480,20 @@ class automated_import_job(osv.osv):
             spamwriter.writerow(pl_row)
         csvfile.close()
 
-        if on_ftp:
+        if on_ftp and job_brw.import_id.ftp_protocol == 'ftp':
             with open(temp_path, 'rb') as temp_file:
                 rep = ftp_connec.storbinary('STOR %s' % pth_filename, temp_file)
                 if not rep.startswith('2'):
-                    raise osv.except_osv(_('Error'), ('Unable to move local file to destination location on FTP server'))
+                    raise osv.except_osv(_('Error'), _('Unable to write report on FTP server'))
+        elif on_ftp and job_brw.import_id.ftp_protocol == 'sftp':
+            new_tmp_file_name = os.path.join(os.path.dirname(temp_path), filename)
+            os.rename(temp_path, new_tmp_file_name)
+            temp_path = new_tmp_file_name
+            try:
+                with sftp.cd(job_brw.import_id.report_path):
+                    sftp.put(new_tmp_file_name, preserve_mtime=True)
+            except:
+                raise osv.except_osv(_('Error'), _('Unable to write report on SFTP server'))
 
         csvfile = open(on_ftp and temp_path or pth_filename, 'r')
         att_obj.create(cr, uid, {
